requirements.txt
paho-mqtt>=1.6.1
Pillow>=10.0.0
requests>=2.31.0
pyyaml>=6.0
python-dotenv>=1.0.0
